requirements.txt
python-dotenv>=1.0
telethon>=1.35
